github_integration_Django/api/serializers.py
from rest_framework import serializers
from .models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["google_id", "github_id", "github_token", "selected_repo"]
