github_integration_Django/requirements.txt
Django==3.2.5  
djangorestframework==3.12.4  
requests==2.25.1  
psycopg2-binary==2.9.3